==> 2 EDA.ipynb <==
Install necessary packages

import numpy as np
import pandas as pd
import pandas_profiling
from pandas_profiling import ProfileReport
import os
os.getcwd()

#Profile report for transformed data

data = pd.read_csv('Data_trans_claims_fac.csv', encoding="ISO-8859-1")
del data["Unnamed: 0"]
data.head()

PR = ProfileReport(data, title="Profile Report - Inspections-Disc")
PR.to_file("Profile_Report-Insp_trans.html")

#Profile report for non-transformed data

data2 = pd.read_csv('Data.csv', encoding="ISO-8859-1")
del data2["Unnamed: 0"]
data2.head()

PR2 = ProfileReport(data2, title="Profile Report - Inspections")
PR2.to_file("Profile_Report-Insp.html")

#Profile report for all inspection data
data3 = pd.read_csv('Data_ALL.csv', encoding="ISO-8859-1")
del data3["Unnamed: 0"]
data3.head()

PR3 = ProfileReport(data3, title="Profile Report - Inspections_ALL")
PR3.to_file("Profile_Report-Insp_ALL.html")
