==> 3 Encoding.ipynb <==
import pandas as pd
import numpy as np
import category_encoders as ce
import sklearn
from sklearn.model_selection import train_test_split

#Import the data
data = pd.read_csv("Data_trans_claims_fac.csv", encoding = "latin-1")
del data["Unnamed: 0"]
data.head()

data.info()

data.isnull().sum()

ENCODING CATEGORICAL VARIABLES

#Claims_level
enc_Claims_Code= ce.OrdinalEncoder(cols=['Claims_Level'], mapping=[{'col':'Claims_Level', 'mapping':{'Very Low':1, 'Low':2, 'Moderate':3, 'High':4, 'Very High':5}}])
data['Claims_Code'] = enc_Claims_code.fit_transform(data[['Claims_Level']])
del data['Claims_Level']

#REGION_CODE
enc_RegCode= ce.OrdinalEncoder(cols=['REGION_CODE'], mapping=[{'col':'REGION_CODE', 'mapping':{'K':1, 'L':2, 'M':3, 'N':4, 'P':5 }}])
data['RegCode'] = enc_RegCode.fit_transform(data[['REGION_CODE']])
del data['REGION_CODE'] 

#IND_Cat
enc_INDCat= ce.OrdinalEncoder(cols=['IND_Cat'], mapping=[{'col':'IND_Cat', 'mapping':{'Goods-producing':1, 'Service-producing':2 }}])
data['INDCat'] = enc_INDCat.fit_transform(data[['IND_Cat']])
del data['IND_Cat'] 

#IND
enc_INDCode= ce.OrdinalEncoder(cols=['IND'], mapping=[{'col':'IND', 'mapping':{'11':1, '21':2, '22':3, '23':4, '31-33':5, '48-49':6, '61':7,'62':8, '72':9, '81':10, '91':11,'Multiple':12 }}])
data['INDCode'] = enc_INDCode.fit_transform(data[['IND']])
del data['IND'] 

#Frac_Level
enc_FrCode= ce.OrdinalEncoder(cols=['Frac_Level'], mapping=[{'col':'Frac_Level', 'mapping':{'Very Low':1, 'Low':2, 'Low-Moderate':3, 'High-Moderate':4, 'High':5, 'Very High':6}}])
data['FRCode'] = enc_FrCode.fit_transform(data[['Frac_Level']])
del data['Frac_Level'] 

#LB_Level
enc_LBCode= ce.OrdinalEncoder(cols=['LB_Level'], mapping=[{'col':'LB_Level', 'mapping':{'Very Low':1, 'Low':2, 'High':3, 'Very High':4}}])
data['LBCode'] = enc_LBCode.fit_transform(data[['LB_Level']])
del data['LB_Level'] 

#Sh_Level
enc_ShCode= ce.OrdinalEncoder(cols=['Sh_Level'], mapping=[{'col':'Sh_Level', 'mapping':{'Very Low':1, 'Low':2, 'High':3, 'Very High':4}}])
data['ShCode'] = enc_ShCode.fit_transform(data[['Sh_Level']])
del data['Sh_Level'] 

#HIC_Level
enc_HICCode= ce.OrdinalEncoder(cols=['HIC_Level'], mapping=[{'col':'HIC_Level', 'mapping':{'Very Low':1, 'Low':2, 'Low-Moderate':3, 'Moderate':4, 'High-Moderate':5, 'High':6, 'Very High':7}}])
data['HICCode'] = enc_HICCode.fit_transform(data[['HIC_Level']])
del data['HIC_Level'] 

#HIC_Per_Level
enc_HICPerCode= ce.OrdinalEncoder(cols=['HIC_Per_Level'], mapping=[{'col':'HIC_Per_Level', 'mapping':{'Very Low':1, 'Low':2, 'Moderate':3, 'High':4, 'Very High':5}}])
data['HICPerCode'] = enc_HICPerCode.fit_transform(data[['HIC_Per_Level']])
del data['HIC_Per_Level'] 

#Emp1524_Level
enc_E1524Code= ce.OrdinalEncoder(cols=['Emp1524_Level'], mapping=[{'col':'Emp1524_Level', 'mapping':{'Below Average':1, 'Above Average':2}}])
data['E1524Code'] = enc_E1524Code.fit_transform(data[['Emp1524_Level']])
del data['Emp1524_Level'] 

#Emp2554_Level
enc_E2554Code= ce.OrdinalEncoder(cols=['Emp2554_Level'], mapping=[{'col':'Emp2554_Level', 'mapping':{'Below Average':1, 'Above Average':2}}])
data['E2554Code'] = enc_E2554Code.fit_transform(data[['Emp2554_Level']])
del data['Emp2554_Level'] 

#Emp55plus_Level
enc_E55pCode= ce.OrdinalEncoder(cols=['Emp55plus_Level'], mapping=[{'col':'Emp55plus_Level', 'mapping':{'Below Average':1, 'Above Average':2}}])
data['E55pCode'] = enc_E55pCode.fit_transform(data[['Emp55plus_Level']])
del data['Emp55plus_Level'] 

#Contra_level
enc_ConCode= ce.OrdinalEncoder(cols=['Contra_level'], mapping=[{'col':'Contra_level', 'mapping':{'None':1, 'Low':2, 'Moderate':3, 'High':4}}])
data['ConCode'] = enc_ConCode.fit_transform(data[['Contra_level']])
del data['Contra_level'] 

list(data.columns)

Split data into 70:30 train / test sets

data.corr()

#Features used --> features with corr >0.85 are removed  --> INDCat, 'FRCode', 'LBCode', 'ShCode',
Features = ['FV_D', 'FV_M', 'FV_DATE', 'VISIT_CODE','WP_ID','NAICS','RegCode', 'INDCode','HICCode','HICPerCode', 'E1524Code', 'E2554Code', 'E55pCode', 'ConCode']
F_x = data.loc[:, Features]
F_y = data.loc[: , ['Claims_Code']]

F_x.corr()

x_tr, x_ts, y_tr, y_ts = train_test_split(F_x, F_y,  train_size = .70, random_state = 23)

#Combine the x and y train set
train = x_tr
train['Claims_Code']= y_tr
train.head()

#Combine the x and y test set
test = x_ts
test['Claims_Code']= y_ts
test.head()

train.to_csv('Train_set_CC.csv')
test.to_csv('Test_set_CC.csv')

list(train.columns)
