==> 4 Feature Selection.ipynb <==
FEATURE SELECTION

import pandas as pd
import numpy as np
import statsmodels.api as sm
import os

# import package scikit-learn 
import sklearn
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import RFECV
from sklearn.svm import SVR
from sklearn.model_selection import StratifiedKFold

import matplotlib.pyplot as plt

os.getcwd()

#Import the data --> this file represents only the entire  set
data = pd.read_csv("Train_set_CC.csv", encoding = "latin-1")
del data["Unnamed: 0"]
data.head()

data.info()

list(data.columns)

data.corr()

# define the data (X) predictors and (y) outcome  

X = data.drop(['FV_DATE','VISIT_CODE', 'Claims_Code'], axis=1)
y = data['Claims_Code']

#split into train and test sets 70:30
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.30, random_state=23)
X.head()

#Univariate chi^2 feature selection
from sklearn.feature_selection import SelectKBest
from sklearn.feature_selection import chi2


#function to select best features using chi^2
def best_feat(X_train, y_train, X_test):
    fs = SelectKBest(score_func=chi2, k="all")
    fs.fit(X_train, y_train)
    X_train_fs = fs.transform(X_train)
    X_test_fs = fs.transform(X_test)
    return X_train_fs, X_test_fs, fs



#run the function
X_tr_feat, X_ts_feat, fs = best_feat(X_train, y_train, X_test)

for i in range(len(fs.scores_)):
    print('Feature %d: %.2f' % (i, fs.scores_[i]))

#4 top important features as per chi^2 approach--> 
X.iloc[:,[2,3,6,5,7]]

# define the data (X) predictors and (y) outcome

X = data.drop(['FV_DATE','VISIT_CODE', 'Claims_Code'], axis=1)
y = data['Claims_Code']

#split into train and test sets 70:30
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.30, random_state=23)



from numpy import mean
from numpy import std
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.feature_selection import RFE
from sklearn.tree import DecisionTreeClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier 
from sklearn.pipeline import Pipeline

#RFE method for identifying most important features
rfe = RFE(estimator=DecisionTreeClassifier(), n_features_to_select=5)
rfe.fit(X_train, y_train)


#Identify most important features using RFE method

from operator import itemgetter
Imp_feat = X_train.columns.to_list()
for x, y in (sorted(zip(rfe.ranking_ , Imp_feat), key=itemgetter(0))):
    print(x, y)
